+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/Kong/kuma
 
-go 1.13
+go 1.14
 
 require (
 	github.com/Kong/kuma/api v0.0.0-00010101000000-000000000000
@@ -31,7 +31,6 @@ require (
 	github.com/prometheus/common v0.9.1
 	github.com/prometheus/prometheus v0.0.0-00010101000000-000000000000
 	github.com/shurcooL/httpfs v0.0.0-20190707220628-8d4bc4ba7749
-	github.com/shurcooL/vfsgen v0.0.0-20181202132449-6a9ea43bcacd // indirect
 	github.com/spf13/cobra v1.0.0
 	github.com/spiffe/go-spiffe v0.0.0-20190820222348-6adcf1eecbcc
 	github.com/spiffe/spire v0.10.0
